Move methods to new class TestRepoCloner
