+++ src/test/java/reposense/git/GitBranchTest.java
@@ -7,6 +7,7 @@ import reposense.model.FileTypeTest;
 import reposense.model.RepoConfiguration;
 import reposense.model.RepoLocation;
 import reposense.template.GitTestTemplate;
+import reposense.util.TestRepoCloner;
 
 
 public class GitBranchTest extends GitTestTemplate {
@@ -25,7 +26,7 @@ public class GitBranchTest extends GitTestTemplate {
         RepoConfiguration uncommonDefaultConfig = new RepoConfiguration(
                 new RepoLocation(TEST_REPO_UNCOMMON_DEFAULT_GIT_LOCATION), RepoConfiguration.DEFAULT_BRANCH);
         uncommonDefaultConfig.setFormats(FileTypeTest.DEFAULT_TEST_FORMATS);
-        GitClone.cloneAndBranch(uncommonDefaultConfig);
+        TestRepoCloner.cloneAndBranch(uncommonDefaultConfig);
         String currentBranch = GitBranch.getCurrentBranch(uncommonDefaultConfig.getRepoRoot());
         Assertions.assertEquals("uncommon", currentBranch);
     }
