+++ src/test/java/reposense/model/RepoConfigurationTest.java
@@ -22,6 +22,7 @@ import reposense.parser.GroupConfigCsvParser;
 import reposense.parser.RepoConfigCsvParser;
 import reposense.report.ReportGenerator;
 import reposense.util.InputBuilder;
+import reposense.util.TestRepoCloner;
 import reposense.util.TestUtil;
 
 public class RepoConfigurationTest {
@@ -126,7 +127,7 @@ public class RepoConfigurationTest {
     @Test
     public void repoConfig_usesStandaloneConfig_success() throws Exception {
         RepoConfiguration actualConfig = new RepoConfiguration(new RepoLocation(TEST_REPO_DELTA), "master");
-        GitClone.cloneAndBranch(actualConfig);
+        TestRepoCloner.cloneAndBranch(actualConfig);
         ReportGenerator.updateRepoConfig(actualConfig);
 
         TestUtil.compareRepoConfig(repoDeltaStandaloneConfig, actualConfig);
@@ -161,7 +162,7 @@ public class RepoConfigurationTest {
         RepoConfiguration.merge(actualConfigs, authorConfigs);
 
         RepoConfiguration actualConfig = actualConfigs.get(0);
-        GitClone.cloneAndBranch(actualConfig);
+        TestRepoCloner.cloneAndBranch(actualConfig);
         ReportGenerator.updateRepoConfig(actualConfig);
 
         TestUtil.compareRepoConfig(expectedConfig, actualConfig);
@@ -184,7 +185,7 @@ public class RepoConfigurationTest {
         RepoConfiguration.setStandaloneConfigIgnoredToRepoConfigs(actualConfigs,
                 cliArguments.isStandaloneConfigIgnored());
         RepoConfiguration actualConfig = actualConfigs.get(0);
-        GitClone.cloneAndBranch(actualConfig);
+        TestRepoCloner.cloneAndBranch(actualConfig);
         ReportGenerator.updateRepoConfig(actualConfig);
 
         TestUtil.compareRepoConfig(expectedConfig, actualConfig);
@@ -212,8 +213,8 @@ public class RepoConfigurationTest {
 
         RepoConfiguration repoBetaActualConfig = actualConfigs.get(0);
         RepoConfiguration repoDeltaActualConfig = actualConfigs.get(1);
-        GitClone.cloneAndBranch(repoBetaActualConfig);
-        GitClone.cloneAndBranch(repoDeltaActualConfig);
+        TestRepoCloner.cloneAndBranch(repoBetaActualConfig);
+        TestRepoCloner.cloneAndBranch(repoDeltaActualConfig);
         ReportGenerator.updateRepoConfig(repoBetaActualConfig);
         ReportGenerator.updateRepoConfig(repoDeltaActualConfig);
         TestUtil.compareRepoConfig(repoBetaExpectedConfig, repoBetaActualConfig);
@@ -236,8 +237,8 @@ public class RepoConfigurationTest {
 
         RepoConfiguration repoBetaActualConfig = actualConfigs.get(0);
         RepoConfiguration repoDeltaActualConfig = actualConfigs.get(1);
-        GitClone.cloneAndBranch(repoBetaActualConfig);
-        GitClone.cloneAndBranch(repoDeltaActualConfig);
+        TestRepoCloner.cloneAndBranch(repoBetaActualConfig);
+        TestRepoCloner.cloneAndBranch(repoDeltaActualConfig);
         ReportGenerator.updateRepoConfig(repoBetaActualConfig);
         ReportGenerator.updateRepoConfig(repoDeltaActualConfig);
         TestUtil.compareRepoConfig(repoBetaExpectedConfig, repoBetaActualConfig);
@@ -256,7 +257,7 @@ public class RepoConfigurationTest {
                 new RepoConfigCsvParser(((ConfigCliArguments) cliArguments).getRepoConfigFilePath()).parse();
 
         RepoConfiguration actualConfig = actualConfigs.get(0);
-        GitClone.cloneAndBranch(actualConfig);
+        TestRepoCloner.cloneAndBranch(actualConfig);
         ReportGenerator.updateRepoConfig(actualConfig);
 
         TestUtil.compareRepoConfig(repoDeltaStandaloneConfig, actualConfig);
@@ -281,7 +282,7 @@ public class RepoConfigurationTest {
         RepoConfiguration.setStandaloneConfigIgnoredToRepoConfigs(actualConfigs, true);
 
         RepoConfiguration actualConfig = actualConfigs.get(0);
-        GitClone.cloneAndBranch(actualConfig);
+        TestRepoCloner.cloneAndBranch(actualConfig);
         ReportGenerator.updateRepoConfig(actualConfig);
 
         TestUtil.compareRepoConfig(expectedConfig, actualConfig);
@@ -307,7 +308,7 @@ public class RepoConfigurationTest {
                 cliArguments.isShallowCloningPerformed());
 
         RepoConfiguration actualConfig = actualConfigs.get(0);
-        GitClone.cloneAndBranch(actualConfig);
+        TestRepoCloner.cloneAndBranch(actualConfig);
         ReportGenerator.updateRepoConfig(actualConfig);
 
         TestUtil.compareRepoConfig(expectedConfig, actualConfig);
@@ -337,8 +338,8 @@ public class RepoConfigurationTest {
 
         RepoConfiguration repoBetaActualConfig = actualConfigs.get(0);
         RepoConfiguration repoDeltaActualConfig = actualConfigs.get(1);
-        GitClone.cloneAndBranch(repoBetaActualConfig);
-        GitClone.cloneAndBranch(repoDeltaActualConfig);
+        TestRepoCloner.cloneAndBranch(repoBetaActualConfig);
+        TestRepoCloner.cloneAndBranch(repoDeltaActualConfig);
         ReportGenerator.updateRepoConfig(repoBetaActualConfig);
         ReportGenerator.updateRepoConfig(repoDeltaActualConfig);
         TestUtil.compareRepoConfig(repoBetaExpectedConfig, repoBetaActualConfig);
@@ -366,8 +367,8 @@ public class RepoConfigurationTest {
 
         RepoConfiguration repoBetaActualConfig = actualConfigs.get(0);
         RepoConfiguration repoDeltaActualConfig = actualConfigs.get(1);
-        GitClone.cloneAndBranch(repoBetaActualConfig);
-        GitClone.cloneAndBranch(repoDeltaActualConfig);
+        TestRepoCloner.cloneAndBranch(repoBetaActualConfig);
+        TestRepoCloner.cloneAndBranch(repoDeltaActualConfig);
         ReportGenerator.updateRepoConfig(repoBetaActualConfig);
         ReportGenerator.updateRepoConfig(repoDeltaActualConfig);
         TestUtil.compareRepoConfig(repoBetaExpectedConfig, repoBetaActualConfig);
@@ -393,7 +394,7 @@ public class RepoConfigurationTest {
         RepoConfiguration.setStandaloneConfigIgnoredToRepoConfigs(actualConfigs, true);
 
         RepoConfiguration actualConfig = actualConfigs.get(0);
-        GitClone.cloneAndBranch(actualConfig);
+        TestRepoCloner.cloneAndBranch(actualConfig);
         ReportGenerator.updateRepoConfig(actualConfig);
 
         TestUtil.compareRepoConfig(expectedConfig, actualConfig);
@@ -419,7 +420,7 @@ public class RepoConfigurationTest {
                 cliArguments.isFindingPreviousAuthorsPerformed());
 
         RepoConfiguration actualConfig = actualConfigs.get(0);
-        GitClone.cloneAndBranch(actualConfig);
+        TestRepoCloner.cloneAndBranch(actualConfig);
         ReportGenerator.updateRepoConfig(actualConfig);
 
         TestUtil.compareRepoConfig(expectedConfig, actualConfig);
@@ -449,8 +450,8 @@ public class RepoConfigurationTest {
 
         RepoConfiguration repoBetaActualConfig = actualConfigs.get(0);
         RepoConfiguration repoDeltaActualConfig = actualConfigs.get(1);
-        GitClone.cloneAndBranch(repoBetaActualConfig);
-        GitClone.cloneAndBranch(repoDeltaActualConfig);
+        TestRepoCloner.cloneAndBranch(repoBetaActualConfig);
+        TestRepoCloner.cloneAndBranch(repoDeltaActualConfig);
         ReportGenerator.updateRepoConfig(repoBetaActualConfig);
         ReportGenerator.updateRepoConfig(repoDeltaActualConfig);
         TestUtil.compareRepoConfig(repoBetaExpectedConfig, repoBetaActualConfig);
@@ -478,8 +479,8 @@ public class RepoConfigurationTest {
 
         RepoConfiguration repoBetaActualConfig = actualConfigs.get(0);
         RepoConfiguration repoDeltaActualConfig = actualConfigs.get(1);
-        GitClone.cloneAndBranch(repoBetaActualConfig);
-        GitClone.cloneAndBranch(repoDeltaActualConfig);
+        TestRepoCloner.cloneAndBranch(repoBetaActualConfig);
+        TestRepoCloner.cloneAndBranch(repoDeltaActualConfig);
         ReportGenerator.updateRepoConfig(repoBetaActualConfig);
         ReportGenerator.updateRepoConfig(repoDeltaActualConfig);
         TestUtil.compareRepoConfig(repoBetaExpectedConfig, repoBetaActualConfig);
@@ -510,8 +511,8 @@ public class RepoConfigurationTest {
 
         RepoConfiguration repoBetaActualConfig = actualConfigs.get(0);
         RepoConfiguration repoDeltaActualConfig = actualConfigs.get(1);
-        GitClone.cloneAndBranch(repoBetaActualConfig);
-        GitClone.cloneAndBranch(repoDeltaActualConfig);
+        TestRepoCloner.cloneAndBranch(repoBetaActualConfig);
+        TestRepoCloner.cloneAndBranch(repoDeltaActualConfig);
         ReportGenerator.updateRepoConfig(repoBetaActualConfig);
         ReportGenerator.updateRepoConfig(repoDeltaActualConfig);
         TestUtil.compareRepoConfig(repoBetaExpectedConfig, repoBetaActualConfig);
@@ -637,7 +638,7 @@ public class RepoConfigurationTest {
                 new RepoConfigCsvParser(((ConfigCliArguments) cliArguments).getRepoConfigFilePath()).parse();
 
         RepoConfiguration actualConfig = actualConfigs.get(0);
-        GitClone.cloneAndBranch(actualConfig);
+        TestRepoCloner.cloneAndBranch(actualConfig);
         ReportGenerator.updateRepoConfig(actualConfig);
 
         TestUtil.compareRepoConfig(expectedConfig, actualConfig);
@@ -659,7 +660,7 @@ public class RepoConfigurationTest {
 
         RepoConfiguration actualConfig = new RepoConfiguration(new RepoLocation(TEST_REPO_MINIMAL_STANDALONE_CONFIG),
                 "master");
-        GitClone.cloneAndBranch(actualConfig);
+        TestRepoCloner.cloneAndBranch(actualConfig);
         ReportGenerator.updateRepoConfig(actualConfig);
 
         TestUtil.compareRepoConfig(expectedConfig, actualConfig);
@@ -695,7 +696,7 @@ public class RepoConfigurationTest {
         RepoConfiguration.merge(actualConfigs, authorConfigs);
 
         RepoConfiguration actualConfig = actualConfigs.get(0);
-        GitClone.cloneAndBranch(actualConfig);
+        TestRepoCloner.cloneAndBranch(actualConfig);
         ReportGenerator.updateRepoConfig(actualConfig);
         Method updateAuthorList = ReportGenerator.class.getDeclaredMethod("updateAuthorList", RepoConfiguration.class);
         updateAuthorList.setAccessible(true);
